# go-architecture/hexagonal-architecture/go-hexagonal-architecture-2/docs/swagger.yaml
basePath: /v1
definitions:
  domain.PaymentType:
    enum:
    - CASH
    - E-WALLET
    - EDC
    type: string
    x-enum-varnames:
    - Cash
    - EWallet
    - EDC
...
  handler.createPaymentRequest:
    properties:
      logo:
        example: https://example.com/cash.png
        type: string
      name:
        example: Tunai
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.PaymentType'
        example: CASH
    required:
    - logo
    - name
    - type
    type: object
...
  handler.errorResponse:
    properties:
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handler.meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  handler.paymentResponse:
    properties:
      id:
        example: 1
        type: integer
      logo:
        example: https://example.com/cash.png
        type: string
      name:
        example: Tunai
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.PaymentType'
        example: CASH
    type: object
...
  handler.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  handler.updatePaymentRequest:
    properties:
      logo:
        example: https://example.com/gopay.png
        type: string
      name:
        example: Gopay
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.PaymentType'
        example: E-WALLET
    required:
    - logo
    - name
    - type
    type: object
...
host: gopos.bagashiz.me
info:
  contact:
    name: test
    url: test
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: MIT
    url: test
  title: Go POS (Point of Sale) API
  version: "1.0"
paths:
...
  /payments:
    get:
      consumes:
      - application/json
      description: List payments with pagination
      parameters:
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payments displayed
          schema:
            $ref: '#/definitions/handler.meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - Payments
    post:
      consumes:
      - application/json
      description: create a new payment with name, type, and logo
      parameters:
      - description: Create payment request
        in: body
        name: createPaymentRequest
        required: true
        schema:
          $ref: '#/definitions/handler.createPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment created
          schema:
            $ref: '#/definitions/handler.paymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - Payments
  /payments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payment deleted
          schema:
            $ref: '#/definitions/handler.response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a payment
      tags:
      - Payments
    get:
      consumes:
      - application/json
      description: get a payment by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payment retrieved
          schema:
            $ref: '#/definitions/handler.paymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a payment
      tags:
      - Payments
    put:
      consumes:
      - application/json
      description: update a payment's name, type, or logo by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payment request
        in: body
        name: updatePaymentRequest
        required: true
        schema:
          $ref: '#/definitions/handler.updatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment updated
          schema:
            $ref: '#/definitions/handler.paymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a payment
      tags:
      - Payments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"